+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - Added `Exxx_ArrayToolsException`
   - Added `E4xx_ArrayToolsException`
   - Added `E4xx_UnsupportedType`
+* Added support for converting key/value pairs in a string into an array
+  - Added `Parsers\ConvertKeyValuePairsToArray`
 
 ## 1.0.0 - Thu 28th Jan 2016